#1328: tests: Make MPI singleton initializie 'on demand' when accessed via Get function
