+++ tests/unit/test_helpers.h
@@ -45,6 +45,7 @@
 #if !defined INCLUDED_TEST_HELPERS_H
 #define INCLUDED_TEST_HELPERS_H
 
+#include "mpi_singleton.h"
 #include "vt/context/context.h"
 #include <gtest/gtest.h>
 
@@ -56,16 +57,12 @@ namespace vt { namespace tests { namespace unit {
  */
 constexpr NodeType CMAKE_DETECTED_MAX_NUM_NODES = vt_detected_max_num_nodes;
 
-
-extern std::unique_ptr<MPISingletonMultiTest> mpi_singleton;
-
 /**
  * Check whether we're oversubscribing on the current execution.
  * This is using MPI because it can be used before vt initializes.
  */
 inline bool isOversubscribed() {
-  assert((mpi_singleton) && "mpi_singleton should be initialized!");
-  return mpi_singleton->getNumRanks() > CMAKE_DETECTED_MAX_NUM_NODES;
+  return MPISingletonMultiTest::Get()->getNumRanks() > CMAKE_DETECTED_MAX_NUM_NODES;
 }
 
 /**
